Completely reworked labeling of caller xrefs
