+++ src/JumpCall.c
@@ -29,6 +29,7 @@
 #include <string.h>
 
 #include "Globals.h"
+#include "Opcodes.h"
 #include "MNemonics.h"
 #include "Tagfile.h"
 #include "JumpCall.h"
@@ -63,6 +64,7 @@ void Register_JumpCall(int From, int To, int Type, unsigned char FunctionCall) {
 		JumpCalls[JumpCall_Count - 1].To = To;
 		JumpCalls[JumpCall_Count - 1].Type = Type;
 		JumpCalls[JumpCall_Count - 1].LabelNumber = 0;
+		JumpCalls[JumpCall_Count - 1].ContainingFunction = -1;
 		JumpCalls[JumpCall_Count - 1].FunctionCall = FunctionCall;
 	}
 }
@@ -106,7 +108,110 @@ void Correct_Label_Types(void) {
 	for (j = LastIdx; j < JumpCall_Count; j++) JumpCalls[j].FunctionCall = CurType;
 }
 
-void Enumerate_Labels(void) {
+int Recurse_Entrypoint(char * Bitstream, int * previous_stack, int depth){
+    int Address = *previous_stack;
+    int recursions_stack[100];
+
+    if(depth > 100){
+        return -1;
+    }
+    recursions_stack[depth] = Address;
+    int i;
+    for(i = 0; i < JumpCall_Count; i++){
+        if(JumpCalls[i].To == Address){
+        }
+    }
+
+    int Opcode_idx_ret = 0;
+    for (i = 0; i < Number_Opcodes; i++) {
+        if (memcmp(MNemonic[Opcodes[i].MNemonic], "ret", 3) == 0) {
+            Opcode_idx_ret = i;
+            break;
+        }
+    }
+
+    int retval = -1;
+    /*printf("Depth: %u\n", depth);*/
+
+    while(Address){
+        /*int Opcode = Get_Next_Opcode(Bitstream + Address);*/
+        int isret = Match_Opcode(Opcodes[Opcode_idx_ret].Opcode_String, Bitstream + Address);
+        int TagIndex;
+        /*printf("At addr %04x, opcode %u\n", Address, Opcode);*/
+        if(isret){
+            /*printf("Is return!\n");*/
+            /*retval = Address+2;*/
+            return -1;
+            /*goto out;*/
+        }
+
+        TagIndex = Tagfile_FindLabelAddress(Address);
+        if(TagIndex != -1){
+            /*printf("Found function %s\n", Tagfile_GetLabel(TagIndex));*/
+            retval = Address;
+            goto out;
+        }
+
+        for(i = 0; i < JumpCall_Count; i++){
+            if(JumpCalls[i].To == Address){
+                if(JumpCalls[i].ContainingFunction >= 0){
+                    /*printf("Found previously known calling function here from %04x!\n", JumpCalls[i].ContainingFunction);*/
+                    return JumpCalls[i].ContainingFunction;
+                }
+                /*printf("Found jump to here from %04x!\n", JumpCalls[i].From);*/
+                recursions_stack[depth+1] = JumpCalls[i].From;
+                /*printf("Recursing...\n");*/
+                int result = Recurse_Entrypoint(Bitstream, recursions_stack+depth+1, depth+1);
+                if(result != -1){
+                    retval = result;
+                    goto out;
+                }
+            }
+            if(JumpCalls[i].From == Address){
+                /*printf("Found jump from here to %04x!\n", JumpCalls[i].To);*/
+                if(JumpCalls[i].ContainingFunction >= 0){
+                    /*printf("Found previously known contained function here from %04x!\n", JumpCalls[i].ContainingFunction);*/
+                    return JumpCalls[i].ContainingFunction;
+                }
+            }
+        }
+        Address -= 2;
+    }
+out: {
+        int j;
+        for(j = 0; j < depth; j++){
+            if(recursions_stack[i] == retval){
+                return -1;
+            }
+        }
+        JumpCalls[i].ContainingFunction = retval;
+        return retval;
+    }
+}
+
+void Label_JumpCalls(char* Bitstream){
+    int i;
+    for(i = 0; i < JumpCall_Count; i++){
+        /*int callee = JumpCalls[i].To;*/
+        /*int TagIndex;*/
+        /*TagIndex = Tagfile_FindLabelAddress(callee);*/
+        if(JumpCalls[i].FunctionCall){
+            /*printf("Function call to function %s at %04x from %04x\n", Tagfile_GetLabel(TagIndex), JumpCalls[i].To, JumpCalls[i].From);*/
+            int caller = Recurse_Entrypoint(Bitstream, &JumpCalls[i].From, 0);
+            JumpCalls[i].ContainingFunction = caller;
+            /*TagIndex = Tagfile_FindLabelAddress(caller);*/
+            /*printf("Label %04x called by function %s at %04x (%04x)\n", JumpCalls[i].To, Tagfile_GetLabel(TagIndex), JumpCalls[i].From, JumpCalls[i].ContainingFunction);*/
+        } else {
+            /*printf("Jump to label at %04x from %04x\n", JumpCalls[i].To, JumpCalls[i].From);*/
+            int caller = Recurse_Entrypoint(Bitstream, &JumpCalls[i].From, 0);
+            JumpCalls[i].ContainingFunction = caller;
+            /*TagIndex = Tagfile_FindLabelAddress(caller);*/
+            /*printf("Label %04x jumped to by function %s at %04x (%04x)\n", JumpCalls[i].To, Tagfile_GetLabel(TagIndex), JumpCalls[i].From, JumpCalls[i].ContainingFunction);*/
+        }
+    }
+}
+
+void Enumerate_Labels(char* Bitstream) {
 	int i;
 	int CurrentLabelNumber = 0;
 	int CurrentFunctionNumber = 0;
@@ -129,6 +234,7 @@ void Enumerate_Labels(void) {
 		if (JumpCalls[i].FunctionCall) JumpCalls[i].LabelNumber = CurrentFunctionNumber;
 			else JumpCalls[i].LabelNumber = CurrentLabelNumber;
 	}
+    Label_JumpCalls(Bitstream);
 }
 
 char *Get_Label_Name(int Destination, char **LabelComment) {
@@ -160,99 +266,9 @@ char *Get_Label_Name(int Destination, char **LabelComment) {
 	return Buffer;
 }
 
-int Resolve_Jump(int Address){
-    int i;
-    for(i = 0; i < JumpCall_Count; i++){
-        // resolve containing label
-		if ((JumpCalls[i].To) <= Address && JumpCalls[i+1].To >= Address) {
-            return i;
-		}
-    }
-    return -1;
-}
-
 char *Get_Containing_Label(int Address){
     char * comment = NULL;
     return Get_Label_Name(Address, &comment);
-}
-
-int Get_Containing_Function(int Address, int depth){
-    int i;
-    if(depth > 100){
-        return -1;
-    }
-    int caller_address = Address;
-    printf("Seeing caller address %04x\n", Address);
-    for(i = 0; i < JumpCall_Count; i++){
-        if ((JumpCalls[i].To) <= caller_address && JumpCalls[i+1].To >= caller_address) {
-
-            int caller = i;
-            printf("Seeing caller label address %04x\n", caller_address);
-
-            int TagIndex;
-            // caller is directly called
-            if(JumpCalls[caller].FunctionCall){
-                printf("Found a direct function call to %04x from %04x\n", JumpCalls[caller].To, JumpCalls[caller].From);
-                /*next_caller_address = JumpCalls[Resolve_Jump(JumpCalls[caller].To)].From;*/
-                TagIndex = Tagfile_FindLabelAddress(caller_address);
-                if (TagIndex != -1) {
-                    printf("Exists\n");
-                    return caller_address;
-                } else {
-                    caller_address = JumpCalls[Resolve_Jump(JumpCalls[caller].From)].To;
-                    printf("Doesn't exist, resolving to %04x\n", caller_address);
-                }
-            }
-
-            // caller address is directly tagged
-            TagIndex = Tagfile_FindLabelAddress(caller_address);
-            if (TagIndex != -1) {
-                printf("Found this caller tag named %s in %04x\n", Tagfile_GetLabel(TagIndex), caller_address);
-                return caller_address;
-            }
-            // caller's caller address is directly tagged
-            int next_caller_address = JumpCalls[Resolve_Jump(caller_address)].From;
-            TagIndex = Tagfile_FindLabelAddress(next_caller_address);
-            if (TagIndex != -1) {
-                printf("Found next caller tag named %s in %04x\n", Tagfile_GetLabel(TagIndex), next_caller_address);
-                return next_caller_address;
-            }
-
-            printf("Resolved as %04x\n", next_caller_address);
-            /*int next_caller_address = JumpCalls[Resolve_Jump(JumpCalls[caller].To)].To;*/
-
-            if(next_caller_address == caller_address){
-                printf("Detected loop\n");
-                /*next_caller_address = JumpCalls[caller-1].From;*/
-                /*next_caller_address = JumpCalls[caller].From;*/
-                /*return Get_Containing_Function(next_caller_address, depth+1);*/
-                /*return caller_address;*/
-                next_caller_address = JumpCalls[Resolve_Jump(caller_address)].From;
-            }
-            int result = -1;
-            if(next_caller_address != caller_address){
-                printf("Tried to resolve %04x to %04x but failed, trying at %04x\n", caller_address, next_caller_address, next_caller_address);
-                result = Get_Containing_Function(next_caller_address, depth+1);
-                printf("Ended up at %04x\n", result);
-            }
-            if(result == -1){
-                printf("End isn't a label, resolving to %04x\n", result);
-                return JumpCalls[Resolve_Jump(next_caller_address)].To;
-            }
-            if(result != -1){
-                TagIndex = Tagfile_FindLabelAddress(result);
-                if(TagIndex == -1){
-                    /*return Get_Containing_Function(result, depth+1);*/
-                    return -1;
-                } else {
-                    return result;
-                }
-            }
-            /*printf("Somethings fucky\n");*/
-            return -1;
-        }
-    }
-    return -1;
 }
 
 /* Show all references which refer to "Position" as destination */
@@ -263,20 +279,16 @@ void Print_JumpCalls(int Position) {
 	for (i = 0; i < JumpCall_Count; i++) {
 		if ((JumpCalls[i].To) == Position) {
 			if (Match == 0) {
-				printf("\n");
+                printf("\n");
 				Match = 1;
 			}
-            // get the index of the JumpCall that is the top-level entrypoint for this target jump
-            int containing_function = Get_Containing_Function(JumpCalls[i].From, 0);
-            if(containing_function > 0){
-                printf("Located jump to %04x in %04x\n", JumpCalls[i].To, containing_function);
+            printf("Container %04x\n", JumpCalls[i].ContainingFunction);
+            const char* caller;
+            if(JumpCalls[i].ContainingFunction >= 0){
+                caller = Get_Containing_Label(JumpCalls[i].ContainingFunction);
+            } else {
+                caller = Get_Containing_Label(JumpCalls[i].From);
             }
-            int call_start = containing_function;
-            if(call_start < 0){
-                /*call_start = JumpCalls[Resolve_Jump(JumpCalls[i].To)].To;*/
-                call_start = JumpCalls[i].From;
-            }
-            const char* caller = Get_Containing_Label(call_start);
             printf("; Referenced from offset 0x%02x <%s> by %s\n", JumpCalls[i].From, caller, MNemonic[JumpCalls[i].Type]);
 		}
 	}
